Move service::error to service::instance::error
This is part of a series of commits to move the existing APIs for
services into a new modules to make room for the new threadless design.

This moves the remaining errors to within the instance module. This
re-exports them from service::instance instead of making the error
module public, unlike previously where it was both re-exported and pub.
